[MoveItem] CAES-1022: Delete old ShareModal Component
